+++ app/policies/project_policy.rb
@@ -26,6 +26,12 @@ class ProjectPolicy
     create?
   end
 
+  def update?
+    return false if @record.completed?
+
+    project_assigned_to_user?
+  end
+
   def openers?
     true
   end
@@ -33,6 +39,10 @@ class ProjectPolicy
   def change_conversion_date?
     return false if @record.conversion_date_provisional?
 
+    project_assigned_to_user?
+  end
+
+  private def project_assigned_to_user?
     @record.assigned_to == @user
   end
 
@@ -52,8 +62,6 @@ class ProjectPolicy
       end
     end
 
-    private
-
-    attr_reader :user, :scope
+    private attr_reader :user, :scope
   end
 end