+++ spec/policies/project_policy_spec.rb
@@ -6,13 +6,32 @@ RSpec.describe ProjectPolicy do
 
   let(:application_user) { build(:user, email: "application.user@education.gov.uk") }
 
+  permissions :update? do
+    it "grants access if project is assigned to the same user" do
+      expect(subject).to permit(application_user, build(:conversion_project, assigned_to: application_user, conversion_date_provisional: false))
+    end
+
+    it "denies access if the project is assigned to another user" do
+      expect(subject).not_to permit(application_user, build(:conversion_project, assigned_to: build(:user), conversion_date_provisional: false))
+    end
+
+    it "denies access if project is assigned to nil" do
+      expect(subject).not_to permit(application_user, build(:conversion_project, assigned_to: nil, conversion_date_provisional: false))
+    end
+
+    it "denies access if the project is completed" do
+      project = build(:conversion_project, assigned_to: application_user, conversion_date_provisional: false, completed_at: Date.yesterday)
+      expect(subject).not_to permit(application_user, project)
+    end
+  end
+
   permissions :change_conversion_date? do
     context "when the conversion date is not provisional" do
       it "grants access if project is assigned to the same user" do
         expect(subject).to permit(application_user, build(:conversion_project, assigned_to: application_user, conversion_date_provisional: false))
       end
 
-      it "denies access if the project is assigend to another user" do
+      it "denies access if the project is assigned to another user" do
         expect(subject).not_to permit(application_user, build(:conversion_project, assigned_to: build(:user), conversion_date_provisional: false))
       end
 
@@ -26,7 +45,7 @@ RSpec.describe ProjectPolicy do
         expect(subject).not_to permit(application_user, build(:conversion_project, assigned_to: application_user, conversion_date_provisional: true))
       end
 
-      it "denies access if the project is assigend to another user" do
+      it "denies access if the project is assigned to another user" do
         expect(subject).not_to permit(application_user, build(:conversion_project, assigned_to: build(:user), conversion_date_provisional: true))
       end